requirement.txt
flask
flask-cors
pyjwt
werkzeug
nmap
python-nmap
pandas
kaggle
